Now adding the original synapse id to the output (#488)
* Now adding the original synapse id to the output

The original synapse id (within CAVE) is now added as an edge property to the output. It can be used to look up all kinds of properties in the original EM data. It is called "cave_id".
Additionally, the "pt_root_id" can now be explicitly provided as input into the task. This way, we do not have to "guess" it from the entity name any more.

* Formatting updates

* Fixed example notebook

* Now converting potential bool data to string, as bool does not work with SONATA.

* Formatting

* Updated the check for bool types

* Changed the test so that no NOQA is required